Sorted functions and type alphabetically.
Function signatures of the (...)Array functions now accept a pointer
to their respective structs, i.e. PDH_FMT_COUNTERVALUE_ITEM_DOUBLE.

The lpdwType parameter of the PdhGetFormattedCounterValue(...) is changed
to a *uint32 instead of the uintptr.

Adjusted examples in the function documentations.

Added a remark about PdhAddEnglishCounterW (which only resides on Windows
Vista and above - need to fix this with a version check in the future).