ui: embed resource the old fashioned way
If we ever get rid of the cgo requirement, we can return to rsrc or some
variant of it. But given that win32 GUI stuff benefits from the larger
cgo stacks, that seems unlikely. This gives us a bit more latitude to
embed all sorts of interesting things in here as well.

Clean up the makefile while we're at it and reduce the size of the exe.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>