tunneltracker: don't track tunnels that haven't been started
Otherwise we get the hasn't-been-started-yet error, and the tracker
quits. Meanwhile this is reported back to the ui as an error. While
we're at it, don't let multiple trackers be run, in the event that the
at-start tracker races with the installation tracker. And, make sure we
actually get the deletion notification.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>