installer: silence ICE61 warning
> ICE61: This product should remove only older versions of itself. The
  Maximum version is not less than the current product.

The warning is caused by <MajorUpgrade AllowSameVersionUpgrades="yes"...

For the time being, this is our design choice. We're shipping WireGuard
with three-field versions 0.0.x. AllowSameVersionUpgrades="yes" avoids
manual configuration of update path in the brain-dead-two-field-version-
comparison-only Group Policy, but requires that the MSI packages are
added to GPO in correct order.

Signed-off-by: Simon Rozman <simon@rozman.si>