installer: silence LGHT1056 warning
> warning LGHT1056 : The Property table contains a row with primary
key(s) 'MsiHiddenProperties' which cannot be merged from the merge
module '.deps\wintun-amd64.msm'.  This is likely due to collision of
rows with the same primary key(s) (but other different values in other
columns) between the database and the merge module.

The warning is caused by Wintun MSM and WireGuard MSI both using
MsiHiddenProperties property. This property is used by WiX to hide
internal data passed to deferred actions from being output to the log.

Signed-off-by: Simon Rozman <simon@rozman.si>