tunnel: do not query real interface name
We want to enforce the relation between the socket name and the service
name, especially since the wintun adapter name might get a 1 or 2
appended depending on weird Nci duplication detection. So we just stick
with the configuration-given name throughout. Also, the user can change
the adapter name at runtime, which is all the more reason why maybe we
shouldn't care about it so much.

Reported-by: Nenad Kozul <me@nenadkozul.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>