ui: fix classic theme drawing with gross hack
Classic theme won't draw transparent images. But new theme erases the
text if we draw ourselves, and we want the OS to draw the text so that
we have better accessibility. Support both by hacking classic theme with
a zero-sized transparent image for the indentation, while using the
transparent image normally for new theme.