tunnel: calculate the actual route metric by summing interface and route metric
I had some issues setting up WireGuard behind another VPN. Curiously, it
bound the physical interface instead of the other VPN, which was the
default route. According to MSDN "the actual route metric used to
compute the route preference is the summation of interface metric
specified in the Metric member of the MIB_IPINTERFACE_ROW structure and
the route metric offset specified in this member" (documentation for
MIB_IPFORWARD_ROW2), but the code did not seem to consider this. After I
changed the calculation, I got the expected behaviour.

Signed-off-by: Philipp Czerner <suyjuris.gi@nicze.de>
Link: https://docs.microsoft.com/en-us/windows/win32/api/netioapi/ns-netioapi-mib_ipforward_row2
[zx2c4: fixed up commit message, removed semicolon]
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>