go: use highres timer on newer Windows 10
WireGuard makes heavy use of timers, and the low precision of Windows
timers as currently used in Go isn't very nice, and also seems to result
in worse battery life because of the gross winmm.dll timerBeginPeriod
trick. John Starks suggested this new parameter to make a high
resolution timer on NT, and Alex implemented it, so import Alex's test
patch to see how this goes.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>