installer: launch GUI before attempting a reinstall
With the recently introduced wireguard-installer.exe an issue in MSI
internal caching appeared. With the temporary MSI file used at install
time gone, the MSI was unable to load our custom actions in the
reinstall attempt.

Rather than attempting to reinstall the product and fail, the MSI was
upgraded to launch GUI early in the reinstall attempt and cancel the
execute sequence then.

Signed-off-by: Simon Rozman <simon@rozman.si>