fetcher: check WinVerifyTrust before execution
Our YubiHSM signature is much stronger than the junky authenticode one,
but still, it can't hurt. This also hedges against anti-virus in the
event that we forget to sign it -- A/V will inspect whatever code the
fetcher executes, and so we only want to execute authenticode-signed
MSIs, to avoid training their heuristics.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>