manager: cleanup legacy wintun
It's not good to leave cruft around from the upgrade case when people
want to uninstall, so we remove wintun in the uninstaller manually
(without the help of wintun.dll, which we no longer ship). But also, so
that we can eventually drop that code, we also remove it
opportunistically when the manager starts.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>