conf: do not examine connectivity state at boot
It turns out that checking for internet connectivity is not really a
reliable way of knowing whether the WSAHOST_NOT_FOUND is legitimate or
not. So just give up on that approach, assume WSAHOST_NOT_FOUND is
always illegitimate at boot, and loop for a long time. This might induce
annoyances for admins who want to kill legitimate WSAHOST_NOT_FOUND
services that keep trying again, but they'll just have to wait for two
minutes.

Reported-by: Simon Rozman <simon@rozman.si>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>