Makefile: choose newest version of Go
Now that these are sorted properly, we can rely on the top one being the
newest. Before, we assumed the last one was the newest, which was only
true during betas and rcs, which required us to swap back and forth
between the first one and the last one. But even then, the last one
wouldn't always be the latest beta! So sometimes we had to change it to
the second to last one. And on and on with madness. In other words,
before the list was basically unsorted. But now that it's sorted, we can
just use the top one always, which is what we want.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>